+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   * 优化 参数构建器的 `page(..)` 与 `limit(..)` 方法，它们起始页码也受页码配置约束
   * 新增 `OrLike` 运算符，参见：https://github.com/ejlchina/bean-searcher/issues/38
   * 方言 `Dialect` 新增 `hasILike()` 方法，当忽略大小写查询时，可利用数据库的索引提升性能
+  * 增强 `EnumFieldConvertor`：支持 `整型` 转换为枚举（按枚举序号转换）
+  * 增强 `EnumFieldConvertor`：新增属性 `failOnError`, 可配置在遇到非法值无法转换时是否报错，默认 `true`
 * Bean Searcher Boot Starter
   * 新增 `bean-searcher.sql.default-mapping.underline-case` 配置键，可配置自动映射时是否开始 驼峰->下划线 的风格转换，默认为 `true`